carrito controlador y item controlador modificados, ResponseDTO creado

+++ pedidos/pedidos/src/main/java/duoc/dairys/pedidos/service/ItemServicio.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import duoc.dairys.pedidos.DTO.ItemCarritoDTO;
 import duoc.dairys.pedidos.model.ItemCarrito;
 import duoc.dairys.pedidos.repository.ItemCarritoRepositorio;
 
@@ -16,13 +17,21 @@ public class ItemServicio {
     private ItemCarritoRepositorio itemCarritoRepositorio;
 
     //crear item
-    public ItemCarrito crear(ItemCarrito item) {
-        if (item.getCantidad() == null || item.getCantidad() <= 0) {
+    public ItemCarrito crear(ItemCarritoDTO dto) {
+        if (dto.getCantidad() == null || dto.getCantidad() <= 0) {
             throw new IllegalArgumentException("Cantidad inválida");
         }
-        if (item.getPrecioUnitario() == null || item.getPrecioUnitario() <= 0) {
+
+        if (dto.getPrecioUnitario() == null || dto.getPrecioUnitario() <= 0) {
             throw new IllegalArgumentException("Precio inválido");
         }
+
+        ItemCarrito item = new ItemCarrito();
+
+        item.setIdProducto(dto.getIdProducto());
+        item.setNombreProducto(dto.getNombreProducto());
+        item.setCantidad(dto.getCantidad());
+        item.setPrecioUnitario(dto.getPrecioUnitario());
 
         item.calcularSubtotal();
 
@@ -36,6 +45,9 @@ public class ItemServicio {
 
     //eliminar
     public void eliminar(Long id) {
+        if (!itemCarritoRepositorio.existsById(id)) {
+            throw new IllegalArgumentException("Item no encontrado");
+        }
         itemCarritoRepositorio.deleteById(id);
     }
